+++ shlink-frontend-kit/src/block/index.ts
@@ -1,3 +1,0 @@
-export * from './Message';
-export * from './Result';
-export * from './SimpleCard';