+++ shlink-frontend-kit/src/form/index.ts
@@ -1,5 +1,0 @@
-export * from './Checkbox';
-export * from './ToggleSwitch';
-export * from './InputFormGroup';
-export * from './LabeledFormGroup';
-export * from './SearchField';