+++ shlink-frontend-kit/src/index.ts
@@ -1,7 +1,0 @@
-export * from './block';
-export * from './form';
-export * from './hooks';
-export * from './navigation';
-export * from './ordering';
-export * from './theme';
-export * from './utils';