+++ shlink-frontend-kit/test/__helpers__/setUpTest.ts
@@ -1,8 +1,0 @@
-import { render } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
-import type { ReactElement } from 'react';
-
-export const renderWithEvents = (element: ReactElement) => ({
-  user: userEvent.setup(),
-  ...render(element),
-});